restir-ish direct lighting works now

+++ shaders/lib/pipelineSettings.glsl
@@ -12,7 +12,7 @@ const int colortex7Format = RGBA16F;		//*cloud map on gbuffers* & temporal filte
 const int colortex8Format = RGBA16F;		//reprojected normal and depth data in prepare
 const int colortex9Format = RGBA32UI;		//scaled depth for atomics in reprojection validation
 const int colortex10Format= RGBA16F;		//block lighting
-const int colortex11Format= RGBA16F;        //valid light sample storage
+const int colortex11Format= RGBA16I;        //valid light sample storage
 */
 
 const bool colortex0Clear = true;
